+++ test/integration/042_sources_test/test_sources.py
@@ -282,6 +282,42 @@ class TestSourceFreshness(SuccessfulSourcesTest):
         # by default, our data set is way out of date!
         self.freshness_start_time = datetime.utcnow()
         results = self.run_dbt_with_vars(
+            ['source', 'freshness', '-o', 'target/error_source.json'],
+            expect_pass=False
+        )
+        self.assertEqual(len(results), 1)
+        self.assertEqual(results[0].status, 'error')
+        self._assert_freshness_results('target/error_source.json', 'error')
+
+        self._set_updated_at_to(timedelta(hours=-12))
+        self.freshness_start_time = datetime.utcnow()
+        results = self.run_dbt_with_vars(
+            ['source', 'freshness', '-o', 'target/warn_source.json'],
+        )
+        self.assertEqual(len(results), 1)
+        self.assertEqual(results[0].status, 'warn')
+        self._assert_freshness_results('target/warn_source.json', 'warn')
+
+        self._set_updated_at_to(timedelta(hours=-2))
+        self.freshness_start_time = datetime.utcnow()
+        results = self.run_dbt_with_vars(
+            ['source', 'freshness', '-o', 'target/pass_source.json'],
+        )
+        self.assertEqual(len(results), 1)
+        self.assertEqual(results[0].status, 'pass')
+        self._assert_freshness_results('target/pass_source.json', 'pass')
+
+    @use_profile('postgres')
+    def test_postgres_source_freshness(self):
+        self._run_source_freshness()
+
+    @use_profile('postgres')
+    def test_postgres_source_snapshot_freshness(self):
+        """Ensures that the deprecated command `source snapshot-freshness`
+        aliases to `source freshness` command.
+        """
+        self.freshness_start_time = datetime.utcnow()
+        results = self.run_dbt_with_vars(
             ['source', 'snapshot-freshness', '-o', 'target/error_source.json'],
             expect_pass=False
         )
@@ -307,10 +343,6 @@ class TestSourceFreshness(SuccessfulSourcesTest):
         self.assertEqual(results[0].status, 'pass')
         self._assert_freshness_results('target/pass_source.json', 'pass')
 
-    @use_profile('postgres')
-    def test_postgres_source_freshness(self):
-        self._run_source_freshness()
-
     @use_profile('snowflake')
     def test_snowflake_source_freshness(self):
         self._run_source_freshness()
@@ -323,6 +355,50 @@ class TestSourceFreshness(SuccessfulSourcesTest):
     def test_bigquery_source_freshness(self):
         self._run_source_freshness()
 
+    @use_profile('postgres')
+    def test_postgres_source_freshness_selection_select(self):
+        """Tests node selection using the --select argument."""
+        self._set_updated_at_to(timedelta(hours=-2))
+        self.freshness_start_time = datetime.utcnow()
+        # select source directly
+        results = self.run_dbt_with_vars(
+            ['source', 'freshness', '--select',
+                'source:test_source.test_table', '-o', 'target/pass_source.json'],
+        )
+        self.assertEqual(len(results), 1)
+        self.assertEqual(results[0].status, 'pass')
+        self._assert_freshness_results('target/pass_source.json', 'pass')
+
+    @use_profile('postgres')
+    def test_postgres_source_freshness_selection_exclude(self):
+        """Tests node selection using the --select argument. It 'excludes' the 
+        only source in the project so it should return no results."""
+        self._set_updated_at_to(timedelta(hours=-2))
+        self.freshness_start_time = datetime.utcnow()
+        # exclude source directly
+        results = self.run_dbt_with_vars(
+            ['source', 'freshness', '--exclude',
+                'source:test_source.test_table', '-o', 'target/exclude_source.json'],
+        )
+        self.assertEqual(len(results), 0)
+
+    @use_profile('postgres')
+    def test_postgres_source_freshness_selection_graph_operation(self):
+        """Tests node selection using the --select argument with graph
+        operations. `+descendant_model` == select all nodes `descendant_model`
+        depends on.
+        """
+        self._set_updated_at_to(timedelta(hours=-2))
+        self.freshness_start_time = datetime.utcnow()
+        # select model ancestors
+        results = self.run_dbt_with_vars(
+            ['source', 'freshness', '--select',
+                '+descendant_model', '-o', 'target/ancestor_source.json']
+        )
+        self.assertEqual(len(results), 1)
+        self.assertEqual(results[0].status, 'pass')
+        self._assert_freshness_results('target/ancestor_source.json', 'pass')
+
 
 class TestSourceFreshnessErrors(SuccessfulSourcesTest):
     @property
@@ -332,7 +408,7 @@ class TestSourceFreshnessErrors(SuccessfulSourcesTest):
     @use_profile('postgres')
     def test_postgres_error(self):
         results = self.run_dbt_with_vars(
-            ['source', 'snapshot-freshness'],
+            ['source', 'freshness'],
             expect_pass=False
         )
         self.assertEqual(len(results), 1)
@@ -348,18 +424,18 @@ class TestSourceFreshnessFilter(SuccessfulSourcesTest):
     def test_postgres_all_records(self):
         # all records are filtered out
         self.run_dbt_with_vars(
-            ['source', 'snapshot-freshness'], expect_pass=False)
+            ['source', 'freshness'], expect_pass=False)
         # we should insert a record with #101 that's fresh, but will still fail
         # because the filter excludes it
         self._set_updated_at_to(timedelta(hours=-2))
         self.run_dbt_with_vars(
-            ['source', 'snapshot-freshness'], expect_pass=False)
+            ['source', 'freshness'], expect_pass=False)
 
         # we should now insert a record with #102 that's fresh, and the filter
         # includes it
         self._set_updated_at_to(timedelta(hours=-2))
         results = self.run_dbt_with_vars(
-            ['source', 'snapshot-freshness'], expect_pass=True)
+            ['source', 'freshness'], expect_pass=True)
 
 
 class TestMalformedSources(BaseSourcesTest):